+++ src/constants/spotify.ts
@@ -1,0 +1,29 @@
+/**
+ * 스포티파이 API 관련 상수
+ */
+
+export const SPOTIFY_API_BASE = "https://api.spotify.com/v1";
+
+/**
+ * 메인 페이지에 표시될 추천 아티스트 ID 목록
+ */
+export const FEATURED_ARTIST_IDS = [
+  "43BxCL6t4c73BQnIJtry5v", // James Hype
+  "7tYKF4w9nC0nq9CsPZTHyP", // SZA
+  "6RHTUrRF63xao58xh9FXYJ", // 아이브
+  "4oLeXFyACqeem2VImYeBFe", // Fred Again..
+  "25uiPmTg16RbhZWAqwLBy5", // Charli XCX
+  "45dkTj5sMRSjrmBSBeiHym", // Tate McRae
+];
+
+/**
+ * 메인 페이지에 표시될 추천 트랙 ID 목록
+ */
+export const RECOMMENDED_TRACK_IDS = [
+  "3sU1L9okYWbN61oHZNQTfh", // Don't Wake Me Up - James Hype
+  "4iZ4pt7kvcaH6Yo8UoZ4s2", // Snooze - SZA
+  "6JjydtOl3rysvM2prpCBaf", // HEYA - 아이브
+  "6NRvZuFXn2ixp8YdzUvG5n", // Leave Me Like This - Skrillex
+  "32VIrOsJmwvqRm4rWFBCsi", // Shiver - John Summit
+  "1tOEEEM8DFyZq3CxSHeq5f", // Kammy (like i do) - Fred again..
+];
