+++ src/features/artist/queries/index.ts
@@ -1,0 +1,4 @@
+export { getArtistAlbums } from "./getArtistAlbums";
+export { getArtistById } from "./getArtistById";
+export { getArtistTopTracks } from "./getArtistTopTracks";
+export { getRelatedArtists } from "./getRelatedArtists";
