+++ src/features/main/queries/index.ts
@@ -1,0 +1,2 @@
+export { getFeaturedArtists } from "./getFeaturedArtists";
+export { getRecommendedTracks } from "./getRecommendedTracks";
