+++ src/features/track/components/TrackInfo.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { SpotifyTrack } from "@/lib/spotify-api";
+import { SpotifyTrack } from "@/types/spotify";
 import { motion } from "framer-motion";
 import Link from "next/link";
 import React from "react";
