+++ src/types/spotify.ts
@@ -1,0 +1,51 @@
+export interface SpotifyImage {
+  url: string;
+  height: number;
+  width: number;
+}
+
+export interface SpotifyArtist {
+  id: string;
+  name: string;
+  images?: SpotifyImage[];
+  genres?: string[];
+  popularity?: number;
+  followers?: {
+    total: number;
+    href: string | null;
+  };
+}
+
+export interface SpotifyAlbum {
+  id: string;
+  name: string;
+  images?: SpotifyImage[];
+  artists: SpotifyArtist[];
+  release_date: string;
+  total_tracks: number;
+  tracks?: {
+    items: SpotifyTrack[];
+  };
+}
+
+export interface SpotifyTrack {
+  id: string;
+  name: string;
+  artists: SpotifyArtist[];
+  album?: SpotifyAlbum;
+  duration_ms: number;
+  preview_url?: string;
+  popularity?: number;
+}
+
+export interface SpotifySearchResult {
+  tracks?: {
+    items: SpotifyTrack[];
+  };
+  artists?: {
+    items: SpotifyArtist[];
+  };
+  albums?: {
+    items: SpotifyAlbum[];
+  };
+}
